% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{analyze_clusters}
\alias{analyze_clusters}
\title{Analyze clusters for size, flattening, likelihood, and point-to-plane distances}
\usage{
analyze_clusters(clusters, ms_gmm_results, covariance)
}
\arguments{
\item{clusters}{cluster dataframe}

\item{ms_gmm_results}{ms_gmm_results dataframe}

\item{covariance}{the covariance dataframe}
}
\value{
vector of log likelihood of each point in a cluster
}
\description{
Analyze clusters for size, flattening, likelihood, and point-to-plane distances
}
