% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_geo.R
\name{cart2sph}
\alias{cart2sph}
\title{Converts from cartesian to spherical coordinates}
\usage{
cart2sph(ned_vec)
}
\arguments{
\item{ned_vec}{this is a vector of direction cosines (i.e. unit vector) in NED coordinates}
}
\value{
an array of the strike and dip (or trend and plunge) in the lower hemispher projection, in radians
}
\description{
Converts from cartesian to spherical coordinates
}
