% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outlier.R
\name{circle3d}
\alias{circle3d}
\title{create a circle around the stage center}
\usage{
circle3d(center = c(0, 0, 0), r = 1, npoints = 100)
}
\arguments{
\item{center}{vector}

\item{r}{radius}

\item{npoints}{int, number of points}
}
\value{
dataframe
}
\description{
create a circle around the stage center
}
