% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_ms.R
\name{classify_ms}
\alias{classify_ms}
\title{Classification algorithm for microseismic events}
\usage{
classify_ms(ms_df, comp_df, comp_tcol = "t", ms_tcol = "t")
}
\arguments{
\item{ms_df}{microseismic data in a dataframe}

\item{comp_df}{completions data}

\item{comp_time_col}{string to identify col of df with start day}

\item{ms_time_col}{string to identify col of df with start day}
}
\description{
Classify events based on time (events must occur after stage start)
to start of stage and proximity to center of stage
Events have to have the same column names, corresponding to x_col,y_col, and z_col
Times should be in posixct
}
