% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{clean_cluster_name}
\alias{clean_cluster_name}
\title{Clean cluster names to match gmm df}
\usage{
clean_cluster_name(df, col)
}
\arguments{
\item{df}{cluster or covariance df}

\item{col}{column to fix}
}
\value{
df with cluster column
}
\description{
Clean cluster names to match gmm df
}
