% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{cluster_log_likelihood}
\alias{cluster_log_likelihood}
\title{Calculate the log likelihood of all points in a cluster}
\usage{
cluster_log_likelihood(this_cluster, ms_gmm_results, covariance)
}
\arguments{
\item{this_cluster}{the current cluster}

\item{ms_gmm_results}{ms_gmm_results}

\item{covariance}{the covariance dataframe}
}
\value{
vector of log likelihood of each point in a cluster
}
\description{
Calculate the log likelihood of all points in a cluster
}
