% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_ms.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix plot}
\usage{
confusion_matrix(ms_df)
}
\arguments{
\item{ms_df}{microseismic data in a dataframe}

\item{comp_df}{completions data}

\item{comp_start_col}{string to identify col of df with start day}
}
\description{
Create and plot a confusion matrix
}
