% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outlier.R
\name{detect_outlier}
\alias{detect_outlier}
\title{Outlier & Induced Event Detection}
\usage{
detect_outlier(ms_df, comp_df, plot = FALSE)
}
\arguments{
\item{ms_df}{microseismic dataframe}

\item{comp_df}{completions dataframe}

\item{plot}{boolean to activate plots or not}

\item{surv_df}{survey dataframe}
}
\description{
Now we investigate each stage using an R-T plot to see if we can remove outliers.
This approach follows Shapiro (2008) and Shapiro (2015), where an apparent
diffusivity is estimated based on the approximation to the PKN model of fracture
growth. See McKean et al. (2019).
}
