% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_geo.R
\name{enu2ned}
\alias{enu2ned}
\title{Converts from ENU (east, north, up) to NED (north, east, down) coordinates}
\usage{
enu2ned(enu_vec)
}
\arguments{
\item{enu_vec}{a vector in ENU coordinates}
}
\value{
a vector in NED coordinates
}
\description{
Converts from ENU (east, north, up) to NED (north, east, down) coordinates
}
