% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_cluster_plot.R
\name{gg_cluster_plot}
\alias{gg_cluster_plot}
\title{Define a ggplot function to plot the data on an x-y, y-z, and z-x views}
\usage{
gg_cluster_plot(ms_df, comp_df, surv_df, well = "1", stage = "1",
  outlier = 0, order = "0", title = TRUE, legend_bool = FALSE,
  label = "", output_path = "../output/")
}
\arguments{
\item{ms_df}{microseismic dataframe with x,y,z}

\item{well}{well label}

\item{stage}{stage label}

\item{outlier}{outlier flag}

\item{order}{order label}

\item{title}{title boolean}

\item{legend_bool}{legend plot boolean}

\item{label}{optional label for filenames}
}
\description{
Define a ggplot function to plot the data on an x-y, y-z, and z-x views
}
