% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_cluster_plot.R
\name{gg_stage_plot}
\alias{gg_stage_plot}
\title{Define a ggplot function to plot the data on an x-y, y-z, and z-x views}
\usage{
gg_stage_plot(ms_df, comp_df, surv_df, well = 1, stage = 1,
  outlier = 0, title = TRUE, legend_bool = FALSE, label = "",
  output_path = "../output/", elev_lim = c(-3000, -2000),
  north_lim = NULL, east_lim = NULL)
}
\arguments{
\item{ms_df}{microseismic dataframe with x,y,z}

\item{well}{well label}

\item{stage}{stage label}

\item{outlier}{outlier flag}

\item{title}{title boolean}

\item{legend_bool}{legend plot boolean}

\item{label}{optional label for filenames}

\item{order}{order label}
}
\description{
Define a ggplot function to plot the data on an x-y, y-z, and z-x views
}
