% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_geo.R
\name{ned2enu}
\alias{ned2enu}
\title{Converts from NED (north, east, down) to ENU (east, north, up) coordinates}
\usage{
ned2enu(ned_vec)
}
\arguments{
\item{ned_vec}{a vector or matrix in NED coordinates, with NED in the columns}
}
\value{
a vector in ENU coordinates
}
\description{
Converts from NED (north, east, down) to ENU (east, north, up) coordinates
}
