% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_ggplot.R
\name{outlier_ggplot}
\alias{outlier_ggplot}
\title{Define a ggplot function to plot the data on an x-y, y-z, and z-x views
Tailored for the the detection outlier aglo}
\usage{
outlier_ggplot(ms_df, comp_df, surv_df, circle_data, well = "1",
  stage = "1", title = TRUE, legend_bool = FALSE)
}
\arguments{
\item{ms_df}{microseismic dataframe with x,y,z,m0}

\item{comp_df}{completions dataframe}

\item{surv_df}{survey dataframe}

\item{circle_data}{shapiro circle data}

\item{well}{well label}

\item{stage}{stage label}

\item{title}{boolean for title}

\item{legend_bool}{boolean for legend}
}
\description{
Define a ggplot function to plot the data on an x-y, y-z, and z-x views
Tailored for the the detection outlier aglo
}
