% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_unit_circle_plot}
\alias{pca_unit_circle_plot}
\title{Calculate and Plot a PCA unit circle}
\usage{
pca_unit_circle_plot(num_df, output_path, colour_vec = NULL)
}
\arguments{
\item{output_path}{output path for plot}

\item{num_ml_df}{clean, numeric only dataframe of ml predictors}
}
\value{
ggplot
}
\description{
Calculate and Plot a PCA unit circle
}
