% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_ggplot.R
\name{plan_ggplot}
\alias{plan_ggplot}
\title{Define a ggplot function to plot the data on an x-y views only
With an inset legend, for publication (no title)}
\usage{
plan_ggplot(ms_df, comp_df, surv_df, well = "3", stage = "2",
  title = "(a)", zoom = FALSE, label = "")
}
\arguments{
\item{ms_df}{microseismic dataframe with x,y,z,m0}

\item{comp_df}{completions dataframe}

\item{surv_df}{survey dataframe}

\item{well}{well label}

\item{stage}{stage label}

\item{title}{char for title string}

\item{zoom}{boolean for zoomed plot}

\item{label}{char to file name label}
}
\description{
Define a ggplot function to plot the data on an x-y views only
With an inset legend, for publication (no title)
}
