% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_geo.R
\name{plane2pole}
\alias{plane2pole}
\title{Converts from the strike/dip of a plane to the trend/plunge of a pole}
\usage{
plane2pole(strike = pi, dip = pi/2)
}
\arguments{
\item{strike}{this is a scalar of the plane strike in radians}

\item{dip}{this is a scalar of the plane dip in radians}
}
\value{
an array of the strike and dip of the plane in radians
}
\description{
Converts from the strike/dip of a plane to the trend/plunge of a pole
}
