% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{point_to_plane_distance_cluster}
\alias{point_to_plane_distance_cluster}
\title{Calculate the distance from points in a cluster to its plane, to be mapped}
\usage{
point_to_plane_distance_cluster(this_cluster, ms_gmm_results, covariance)
}
\arguments{
\item{this_cluster}{the current cluster}

\item{ms_gmm_results}{ms_gmm_results}

\item{covariance}{the covariance dataframe}
}
\value{
vector of distances between each point in a cluster and its plane
}
\description{
Calculate the distance from points in a cluster to its plane, to be mapped
}
