% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_geo.R
\name{pole2plane}
\alias{pole2plane}
\title{Converts from the trend/plunge of a pole to the strike/dip of a plane}
\usage{
pole2plane(trend = pi, plunge = pi/2)
}
\arguments{
\item{trend}{this is a scalar of the pole trend in radians}

\item{plunge}{this is a scalar of the pole plunge in radians}
}
\value{
an array of the strike and dip of the plane in radians
}
\description{
Converts from the trend/plunge of a pole to the strike/dip of a plane
}
