% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ms.R
\name{process_ms}
\alias{process_ms}
\title{Function to process raw microseismic data into a clean data.frame for subsequent analysis}
\usage{
process_ms(ms_df, mag_completness = -2, loc_cols = c("x", "y", "z"),
  date_time_cols = c("date", "time"), mag_col = c("mw"),
  feat_cols = NA, crs = 26911, tz = "UTC")
}
\arguments{
\item{ms_df}{raw microseismic dataframe}

\item{crs}{projection of the data}

\item{mag_complete}{moment magnitude of completeness of the dataset}

\item{ms_feat}{a list of features to carry into the cluster analysis}

\item{names}{of required columns - character vector of time, x, y, z, and magnitude}
}
\value{
data.frame with t,x,y,z,and any other features
}
\description{
function takes a raw ms dataframe with projected coordinates (e.g. UTM NAD 83)
and processes an analysi
}
