% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_comp.R
\name{resample_comp_df}
\alias{resample_comp_df}
\title{Function to resample location of hydraulic fracturing}
\usage{
resample_comp_df(comp_df, survey_df)
}
\arguments{
\item{comp_df}{completions dataframe}

\item{survey_df}{well survey}
}
\value{
resampled completions data frame
}
\description{
resamples completions dataframe with uniform distribution
}
