% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gmm.R
\name{run_gmm}
\alias{run_gmm}
\title{Run the Gaussian Mixture Model function}
\usage{
run_gmm(ms_outlier, comp_df, surv_df, comp_start_col = "t",
  ms_feat = c("x", "y", "z", "t"), ms_time_col = "t",
  title_bool = FALSE, legend_bool = FALSE, plot = TRUE, ...)
}
\arguments{
\item{ms_outlier}{microseismic data in a dataframe}

\item{comp_df}{completions data}

\item{surv_df}{survey day}

\item{comp_start_col}{string to identify col of df with start day}

\item{ms_feat}{a list of columns for the clustering}
}
\description{
Run the MClust package with up to 30 clusters and a VVV model
}
