% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gmm_parallel.R
\name{run_gmm_parallel}
\alias{run_gmm_parallel}
\title{Run the Gaussian Mixture Model function}
\usage{
run_gmm_parallel(ms_df, comp_df, surv_df, run = 1,
  comp_start_col = "t", ms_feat = c("x", "y", "z", "t"),
  ms_time_col = "t", title_bool = FALSE, legend_bool = FALSE,
  plot = FALSE, ...)
}
\arguments{
\item{ms_df}{microseismic data in a dataframe}

\item{comp_df}{completions data}

\item{surv_df}{survey day}

\item{run}{run number}

\item{comp_start_col}{string to identify col of df with start day}

\item{ms_feat}{a list of columns for the clustering}
}
\description{
Run the MClust package with up to 30 clusters and a VVV model
}
