% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_geo.R
\name{sph2cart}
\alias{sph2cart}
\title{Converts from spherical to cartesian coordinates}
\usage{
sph2cart(strike_or_trend = pi, dip_or_plunge = pi/2, plane = TRUE)
}
\arguments{
\item{strike_or_trend}{this is a scalar of the trend or strike in radians}

\item{dip_or_plunge}{this is a scalar of the plunge or dip in radians}

\item{plane}{this is a boolean that selects whether the strike/dip of a plane (TRUE) or trend/plunge of a line (FALSE) are analyzed}
}
\value{
an array of the north,east,down (NED) direction cosines (i.e. unit vector)
}
\description{
Converts from spherical to cartesian coordinates
}
